Only seed product details JSON once

+++ scripts/seed-prod-details.sh
@@ -9,10 +9,14 @@ if [ ! -d $dest ]; then
     echo "you must run this from the root of your zamboni checkout"
     exit 1
 fi
-svn export --force http://svn.mozilla.org/libs/product-details/json/ $dest
-if [ $? -eq 0 ]; then
-    # Bah. This isn't supported for some reason: date "+%a, %d %b %Y %H:%M:%S %Z"
-    dd=`python -c 'import datetime,sys; sys.stdout.write(datetime.datetime.utcnow().strftime("%a, %d %b %Y %H:%M:%S GMT"))'`
-    echo -n $dd > $dest/.last_update
-    echo -n $dd > $dest/regions/.last_update
+if [ ! -f $dest/.last_update ]; then
+    svn export --force http://svn.mozilla.org/libs/product-details/json/ $dest
+    if [ $? -eq 0 ]; then
+        # Bah. This isn't supported for some reason: date "+%a, %d %b %Y %H:%M:%S %Z"
+        dd=`python -c 'import datetime,sys; sys.stdout.write(datetime.datetime.utcnow().strftime("%a, %d %b %Y %H:%M:%S GMT"))'`
+        echo -n $dd > $dest/.last_update
+        echo -n $dd > $dest/regions/.last_update
+    fi
+else
+    echo "Already seeded product details JSON"
 fi
